refactor(theme): better data handle + comments
